requirements.txt
python-telegram-bot==20.7
tinydb==4.8.0
python-dotenv==1.0.0
httpx==0.25.2
